requirements_dev.txt
colossus>=1.2.1
coverage==4.5.1
coveralls
cryptography==2.1.4
flake8==2.6.0
mpmath==1.0.0
nbsphinx
numpydoc==0.8.0
pip
pytest==3.5.1
pytest-cov==2.5.1
pytest-runner==4.2
PyYAML==3.12
Sphinx==1.7.5
sphinx_rtd_theme
sympy==1.1.1
tox==3.0.0
wheel==0.31.0
